[Bronze V] Title: 이상한 기호, Time: 104 ms, Memory: 14160 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -2,7 +2,7 @@ import java.util.*;
 import java.io.*;
 
 public class Main {
-    public static int func(int A, int B) {
+    public static long func(long A, long B) {
         return (A + B) * (A - B);
     }
 
@@ -10,8 +10,8 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         String str = br.readLine();
         StringTokenizer st = new StringTokenizer(str, " ");
-        int A = Integer.parseInt(st.nextToken());
-        int B = Integer.parseInt(st.nextToken());
+        long A = Integer.parseInt(st.nextToken());
+        long B = Integer.parseInt(st.nextToken());
         System.out.println(func(A, B));
     }
 }
